LayoutTests/fast/forms/date-multiple-fields/date-multiple-fields-wheel-event-expected.txt

Test for wheel operations for <input type=date>

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Initial value is 1999-01-13. We'll wheel up by 1:
PASS input.value is "1999-02-13"
Wheel up by 100:
PASS input.value is "1999-03-13"
Wheel down by 1:
PASS input.value is "1999-02-13"
Wheel down by 256:
PASS input.value is "1999-01-13"
Disabled input element:
PASS input.value is "1999-01-13"
Read-only input element:
PASS input.value is "1999-01-13"
No focus:
PASS input.value is "1999-01-13"
PASS successfullyParsed is true

TEST COMPLETE
